Create Problem3.js
I was still having problems with the actual P5.js program not running. I checked the syntax and rechecked the logic. Unfortunately I don't know the developer tools to fix the problem with the Chrome browser on my own so have no way of testing this. It's sort of a problem, and after going into console, it only displays the problem 1/4 of the time.  If there are any alternatives to p5, or a solid get around for this happening all the time, it would be very useful to know. Thank you.